FST: Fix reading of the network speed value

+++ fst/ScanDir.cc
@@ -1070,7 +1070,7 @@ ScanDir::IsValidStripeCombination(
     return false;
   }
 
-  uint32_t offsetXrd = 0;
+  uint64_t offsetXrd = 0;
   const auto open_ts = std::chrono::system_clock::now();
   int scan_rate = mRateBandwidth.load(std::memory_order_relaxed);
   xs_obj->Reset();
